Updated tests for changes in implementations
